Manejo de errores y exepciones

+++ src/main/java/uy/edu/ucu/inventario/service/DepositoService.java
@@ -5,6 +5,8 @@ import uy.edu.ucu.inventario.repository.DepositoRepository;
 
 import org.springframework.stereotype.Service;
 
+import jakarta.persistence.EntityNotFoundException;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -33,6 +35,10 @@ public class DepositoService {
     }
 
     public void eliminar(Long id) {
+
+        if (!repo.existsById(id)) {
+            throw new EntityNotFoundException("Depósito con id " + id + " no encontrado");
+        }
         repo.deleteById(id);
     }
 }
